// API/src/Cards/CardsEnum.ts
export enum CardTypes {
    SPELL = "SPELL",
    ENCHANTMENT = "ENCHANTMENT",
    EKIP = "EKIP",
    CONTRACT = "CONTRACT",
    CHARACTER = "CHARACTER",
}

export enum CardActions {
    TAKE = "TAKE",
    DESTROY = "DESTROY",
    ADD = "ADD",
    DRAW = "DRAW",
    CHANGE = "CHANGE",
    RENEW = "RENEW",
}

export enum CardPlaces {
    DECK = "DECK",
    HAND = "HAND",
    DUMP = "DUMP",
    SHOP = "SHOP",
}

export enum CardResources {
    COIN = "COIN",
    STRENGTH = "STRENGTH",
    TOUGHNESS = "TOUGHNESS",
    CARD = "CARD",
    REWARD = "REWARD",
}

export enum EkipElements {
    FIRE = "FIRE",
    WATER = "WATER",
    EARTH = "EARTH",
    AIR = "AIR"
}

export enum EkipParts {
    HEAD = "HEAD",
    BODY = "BODY",
    LEGS = "LEGS",
    ARMS = "ARMS"
}